Updates to the model structure and get_scoped_attributes functions. 1: Split out nodes, links, groups and network into separate model files as the network file was getting too long 2: Add get_scoped_attributes to the network and Project classes to fetch attribtues scoped to them, and to look up the hierarchy if requested. Simplify the calling code as appropriate.
