+++ hydra_base/db/model/network/__init__.py
@@ -1,2 +1,5 @@
-from .network import *
 from .resourceattr import *
+from .node import *
+from .link import *
+from .resourcegroup import *
+from .network import *
